removing formalize from the Admin theme
The buttons were being restyled with formalize, and that was messing up the ability to use them. I see no need to override Seven's button style. I believe this was added in an attempt to style content entry forms for content that will actually be styled by the Food Coop theme. So I'm removing Formalize entirely from the Admin theme.