Adding Timefield Module
I'm building a field collection for the new Drupal site to have a data model for storing squad information. I need a field for the timeslot. Date 7.x-2.x can't store a time without a year. Apparently this will be added to 7.x-3.x, but that branch is not ready. So I'm trying this module out.