foodcoop patch to contrib module - fix for one time login link
This issue has a seeming resemblance to this issue:

https://www.drupal.org/node/2455049

But that issue was resolved and the problem persists.

To test:

Go through coop process to create a new account
Recieve email with password reset URL
Click on link in email
Get error that your hash has been used or is wrong.